skills/tts/pyproject.toml
[project]
name = "tts"
version = "0.1.0"
dependencies = [
    "requests",
    "python-dotenv",
    "google-genai",
]

[project.scripts]
tts = "scripts.tts:main"
